Add progress bar to bundle
