+++ Cargo.toml
@@ -39,6 +39,7 @@ tar = "0.4.42"
 zip = "2.2.0"
 zstd = "0.13.2"
 bzip2 = "0.4.4"
+indicatif = "0.17.8"
 
 [dev-dependencies]
 assert_cmd = "2.0.14"
